# openapi.yaml
openapi: 3.0.3
info:
  title: E-commerce Graphic Designer API
  description: AI-powered image editing service for e-commerce product images
  version: 1.0.0
  contact:
    name: Development Team
    email: dev@example.com
servers:
  - url: http://localhost:3001/api
    description: Development server
  - url: https://api.example.com/api
    description: Production server
paths:
...
  /job/{id}:
    get:
      tags: ['Jobs']
      summary: Get job status
      description: Retrieve current status and results of a processing job
      parameters:
        - name: id
          in: path
          required: true
          description: Job ID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Job status information
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/JobStatus'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /job/stream/{id}:
    get:
      tags: ['Jobs']
      summary: Stream job progress
      description: Real-time streaming of job progress using Server-Sent Events
      parameters:
        - name: id
          in: path
          required: true
          description: Job ID
          schema:
            type: string
            format: uuid
        - name: Last-Event-ID
          in: header
          required: false
          description: Last received event ID for reconnection
          schema:
            type: string
      responses:
        '200':
          description: Event stream
          content:
            text/event-stream:
              schema:
                type: string
                description: Server-sent events with progress updates
              examples:
                progress:
                  value: "event: progress\ndata: {\"status\":\"running\",\"progress\":50,\"error_msg\":null}\n\n"
                complete:
                  value: "event: complete\ndata: {\"job_id\":\"123e4567-e89b-12d3-a456-426614174000\"}\n\n"
                error:
                  value: "event: error\ndata: {\"error\":\"Processing failed\"}\n\n"
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  schemas:
    ApiResponse:
      type: object
      required:
        - success
      properties:
        success:
          type: boolean
          description: Whether the request was successful
        data:
          type: object
          nullable: true
          description: Response data (present on success)
        error:
          type: string
          nullable: true
          description: Error message (present on failure)
    AIModelType:
      type: string
      enum: [gemini, chatgpt, sora]
      description: Available AI models
...
    Variant:
      type: object
      properties:
        id:
          type: string
          format: uuid
        job_id:
          type: string
          format: uuid
        image_id:
          type: string
          format: uuid
        score:
          type: number
          minimum: 0
          maximum: 1
          description: Quality score of the variant
        thumbnail_path:
          type: string
          description: Path to thumbnail image
        metadata:
          type: object
          additionalProperties: true
          description: Variant metadata including AI model used
    JobStatus:
      type: object
      properties:
        id:
          type: string
          format: uuid
        status:
          type: string
          enum: [pending, queued, running, done, error, failed]
          description: Current job status
        progress:
          type: integer
          minimum: 0
          maximum: 100
          description: Processing progress percentage
        result_variants:
          type: array
          items:
            $ref: '#/components/schemas/Variant'
          description: Generated variants (available when status is 'done')
        error_msg:
          type: string
          nullable: true
          description: Error message if job failed
        model_used:
          $ref: '#/components/schemas/AIModelType'
          nullable: true
          description: AI model used for processing
        attempts:
          type: integer
          minimum: 1
          description: Number of processing attempts
        last_error:
          type: string
          nullable: true
          description: Last error encountered during processing
...
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/ApiResponse'
              - type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  error:
                    type: string
                    example: "Resource not found"
...
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/ApiResponse'
              - type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  error:
                    type: string
                    example: "Internal server error"
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for authentication (if implemented)
...